+++ pyramid_swagger/tween.py
@@ -21,6 +21,7 @@ from pyramid.interfaces import IRoutesMapper
 from pyramid.settings import asbool
 from pyramid.settings import aslist
 
+from pyramid_swagger.exceptions import PathNotFoundError
 from pyramid_swagger.exceptions import RequestValidationError
 from pyramid_swagger.exceptions import ResponseValidationError
 from pyramid_swagger.model import PathNotMatchedError
@@ -167,15 +168,16 @@ def validation_tween_factory(handler, registry):
         except PathNotMatchedError as exc:
             if settings.validate_path:
                 with validation_context(request):
-                    raise RequestValidationError(str(exc))
+                    raise PathNotFoundError(str(exc), child=exc)
             else:
                 return handler(request)
 
         if settings.validate_request:
-            request_data = swagger_handler.handle_request(
-                PyramidSwaggerRequest(request, route_info),
-                op_or_validators_map,
-                validation_context=validation_context)
+            with validation_context(request, response=None):
+                request_data = swagger_handler.handle_request(
+                    PyramidSwaggerRequest(request, route_info),
+                    op_or_validators_map,
+                )
 
             def swagger_data(_):
                 return request_data
@@ -310,7 +312,7 @@ class PyramidSwaggerResponse(OutgoingResponse):
         return getattr(self.response, 'json_body', {})
 
 
-def handle_request(request, validator_map, validation_context, **kwargs):
+def handle_request(request, validator_map, **kwargs):
     """Validate the request against the swagger spec and return a dict with
     all parameter values available in the request, casted to the expected
     python type.
@@ -318,10 +320,10 @@ def handle_request(request, validator_map, validation_context, **kwargs):
     :param request: a :class:`PyramidSwaggerRequest` to validate
     :param validator_map: a :class:`pyramid_swagger.load_schema.ValidatorMap`
         used to validate the request
-    :param validation_context: a context manager for wrapping validation
-        errors
     :returns: a :class:`dict` of request data for each parameter in the swagger
         spec
+    :raises: RequestValidationError when the request is not valid for the
+        swagger spec
     """
     request_data = {}
     validation_pairs = []
@@ -343,8 +345,7 @@ def handle_request(request, validator_map, validation_context, **kwargs):
         validation_pairs.append((validator_map.body, request.body))
         request_data[param_name] = request.body
 
-    with validation_context(request):
-        validate_request(validation_pairs)
+    validate_request(validation_pairs)
 
     return request_data
 
@@ -468,7 +469,7 @@ def validation_error(exc_class):
                 # This will alter our stack trace slightly, but Pyramid knows
                 # how to render it. And the real value is in the message
                 # anyway.
-                e = exc_class(str(exc))
+                e = exc_class(str(exc), child=exc)
                 e._traceback = sys.exc_info()[2]
                 raise e
 
@@ -569,11 +570,8 @@ def swaggerize_request(request, op, **kwargs):
 
     :type request: :class:`pyramid.request.Request`
     :type op: :class:`bravado_core.operation.Operation`
-    :type validation_context: context manager
     """
-    validation_context = kwargs['validation_context']
-    with validation_context(request):
-        request_data = unmarshal_request(request, op)
+    request_data = unmarshal_request(request, op)
     return request_data